+++ ReduxObservable.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
