+++ ReduxObservable/ReduxAction.swift
@@ -1,0 +1,13 @@
+//
+//  ReduxAction.swift
+//  ReduxExamples
+//
+//  Created by Camilo Ortegon on 10/9/19.
+//  Copyright © 2019 Grability. All rights reserved.
+//
+
+import Foundation
+
+public protocol ReduxAction {
+    var identifier: String { get }
+}
