+++ ReduxObservable/ReduxObservable.h
@@ -1,0 +1,19 @@
+//
+//  ReduxObservable.h
+//  ReduxObservable
+//
+//  Created by daniel parra on 10/27/19.
+//  Copyright © 2019 grability. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for ReduxObservable.
+FOUNDATION_EXPORT double ReduxObservableVersionNumber;
+
+//! Project version string for ReduxObservable.
+FOUNDATION_EXPORT const unsigned char ReduxObservableVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <ReduxObservable/PublicHeader.h>
+
+
