+++ ReduxObservable/Rx+Operators.swift
