+++ ReduxObservableTests/middlewares/EventRecorderMiddleware.swift
@@ -1,0 +1,25 @@
+//
+//  EventCounterMiddleware.swift
+//  redux-observableTests
+//
+//  Created by Camilo Ortegon on 10/10/19.
+//  Copyright © 2019 Grability. All rights reserved.
+//
+
+import RxSwift
+@testable import ReduxObservable
+
+class EventRecorderMiddleware: EventCounterMiddleware {
+
+    var actions = [ReduxAction]()
+
+    init() {
+        super.init("")
+    }
+
+    override func dispatch(store: Store<TestAppState>, action: ReduxAction) {
+        actions.append(action)
+        actionRelay.onNext(action.identifier)
+    }
+
+}
