--- redux-observable/ViewController.swift
@@ -1,17 +1,0 @@
-//
-//  ViewController.swift
-//  redux-observable
-//
-//  Created by Camilo Ortegon on 10/10/19.
-//  Copyright © 2019 Grability. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-
-}
