--- redux-observable/rx-redux/library/Middleware.swift
@@ -1,14 +1,0 @@
-//
-//  Middleware.swift
-//  ReduxExamples
-//
-//  Created by Camilo Ortegon on 10/9/19.
-//  Copyright © 2019 Grability. All rights reserved.
-//
-
-class Middleware<S: Equatable> {
-
-    func dispatch(store: Store<S>, action: ReduxAction) {
-    }
-
-}
