--- redux-observable/rx-redux/library/Reducer.swift
@@ -1,13 +1,0 @@
-//
-//  Reducer.swift
-//  ReduxExamples
-//
-//  Created by Camilo Ortegon on 10/9/19.
-//  Copyright © 2019 Grability. All rights reserved.
-//
-
-class Reducer<S> {
-    func reduce(state: S, action: ReduxAction) -> S {
-        return state
-    }
-}
