--- redux-observable/rx-redux/library/ReduxAction.swift
@@ -1,13 +1,0 @@
-//
-//  ReduxAction.swift
-//  ReduxExamples
-//
-//  Created by Camilo Ortegon on 10/9/19.
-//  Copyright © 2019 Grability. All rights reserved.
-//
-
-import Foundation
-
-protocol ReduxAction {
-    var identifier: String { get }
-}
